Models removed from the models folder and assigned to the corresponding DAOs, as well as making the connection to MongoDB implicit.
