+++ dao/chatsDAO.js
@@ -1,10 +1,32 @@
-const {chats} = require ('../models');
+const mongoose = require ('mongoose')
+const config = requite ('../config/config.json')
+const chatSchema = mongoose.Schema({
+    participants: [{
+        username: {
+          type:String
+        }
+      }],
+      messages: [{
+        username:{
+          type:String
+        },
+        message:{
+          type:String
+        },
+        dateOfMessage:{
+          type:Date
+        }
+      }]
+});
+const chats = mongoose.model('chats', chatSchema);
+mongoose.connect(config.development.database.url);
+
 class chatsDAO{
     static async startNewChat(newChat){
         const chatExistent = await chats.findOne(newChat.participants);
         if (!chatExistent) {
             const chat = new chats({ participants, messages });
-            return await chats.create(chat);
+            return await chats.save(chat);
         }
     }
     static async addMessageToChat(newChat) {
