+++ dao/postsDAO.js
@@ -1,4 +1,114 @@
-const {posts} = require ('../models');
+'use strict';
+const mongoose = require ('mongoose');
+const config = require('../config/config.json')
+const postSchema = mongoose.Schema({
+  id:{
+    type: Number,
+    unique: true
+  },
+  title: {
+    type: String
+  },
+  author: {
+    type: String
+  },
+  body: {
+    type: String
+  },
+  dateOfPublish: {
+    type: Date
+  },
+  photos: [{
+    url: {
+      type: String,
+      required: false
+    }
+  }],
+  target: {
+    type: String
+  },
+  comments: [{
+    commentId:{
+      type: Number,
+      unique: true,
+      required: false
+    },
+    author: {
+      type: String,
+      required: false
+    },
+    body: {
+      type: String,
+      required: false
+    },
+    dateOfComment: {
+      type: Date,
+      required: false
+    }
+  }],
+  likes: {
+    type: Number,
+    required: false
+  },
+  dislikes: {
+    type: Number,
+    required: false
+  },
+  reports: {
+    type: Number,
+    required: false
+  }
+});
+
+const posts = mongoose.model('posts', postSchema)
+//Method to auto-increment the post id
+mongoose.connect(config.development.database.url);
+postSchema.pre('create', function (next) {
+  const doc = this;
+  if (doc.isNew) {
+    mongoose.model('posts', postSchema)
+    .findOne({}, {}, { sort: { 'id': -1 } })
+    .exec()
+    .then((lastDocument) => {
+      if (lastDocument) {
+        doc.id = lastDocument.id + 1;
+      } else {
+        doc.id = 1;
+      }
+      next();
+    })
+    .catch((err) => {
+      next(err);
+    });
+  } else {
+    next();
+  }
+});
+//Method to auto-increment the comment id for each post
+postSchema.pre('create', function (next) {
+  const doc = this;
+  doc.comments.forEach(comment => {
+    if (comment.isNew) {
+      mongoose.model('posts', postSchema)
+      .findOne({}, {}, { sort: { 'comments.commentId': -1 } })
+      .exec()
+      .then((lastDocument) => {
+        let lastCommentID = 1;
+        if (lastDocument && lastDocument.comments.length > 0) {
+          const lastComment = lastDocument.comments.find(c => c.commentId);
+          if (lastComment) {
+            lastCommentID = lastComment.commentId + 1;
+          }
+        }
+        comment.commentId = lastCommentID;
+      })
+      .catch((err) => {
+        next(err);
+      });
+    }
+  });
+  next();
+});
 class postsDAO{
     static async addNewPost(post){
         return await posts.create(post);
